requirements.txt
apify_client==1.0.0
cleantext==1.1.4
matplotlib==3.5.0
nltk==3.8.1
numpy==1.23.2
pandas==1.3.4
Pillow==9.5.0
plotly==5.10.0
requests==2.27.1
streamlit==1.21.0
textblob==0.17.1
